Adds --start-at-preview and --stop-at-duration and --stop-at-pts options. These correspond to the job->start_at_preview and job->pts_to_stop variables the MacGui and LinGui use to achieve live preview encodes.
While of course starting at a certain preview is not an optimal way of choosing a starting position for an encode (especially in a CLI), it is fast, and flexible. By altering the number of previews to scan with --previews, it is possible to start an encode at any percentage of the video.

git-svn-id: svn://localhost/HandBrake/trunk@2051 b64f7644-9d1e-0410-96f1-a4d463321fa5