Fix mkv timestamps that derive from SimpleBlock and default track duration. Since I've gotten no response to this patch yet on ffmpeg-devel, I'm committing it here till someone has the time to look at it.
When a SimpleBlock has lacing (multiple frames in a block) the duration
of each frame should come from the default track duration.  An initialization
error in ffmpeg causes them to set the duration to 0 which triggers
code higher in the stack to estimate the duration based on bitrate
and frame size.  This is wildly inaccurate and leads to very bad timestamps.

Fixes issues with MakeMKV audio timestamps.


git-svn-id: svn://localhost/HandBrake/trunk@3835 b64f7644-9d1e-0410-96f1-a4d463321fa5